feat: generate readme and demo

+++ src/cli.js
@@ -3,7 +3,7 @@
 const fs = require('fs')
 const path = require('path')
 const execa = require('execa')
-const testFiles = ['babel.config.js', 'index.test.js', 'jest.config.js', 'rollup.config.js', 'rollup.config.split.js']
+const testFiles = ['babel.config.js', 'index.test.js', 'jest.config.js', 'rollup.config.js', 'rollup.config.split.js', 'generate.js', 'index.html']
 
 const copyTestFiles = () => {
   testFiles.forEach((file) => {
@@ -13,49 +13,60 @@ const copyTestFiles = () => {
 
 const removeTestFiles = () => {
   testFiles.forEach((file) => {
-    fs.rmSync(file)
+    if (file !== 'index.html') {
+      fs.rmSync(file)
+    }
   })
 }
 
-const onError = () => {}
+const onError = () => {
+  removeTestFiles()
+}
 
-execa(
-  path.join('.', 'node_modules', '.bin', 'eslint') +
-  ' --fix "**/*.*" --resolve-plugins-relative-to . --ignore-path ' +
-  path.join('.', 'node_modules', 'web-animations-set', 'src', '.eslintignore') +
-  ' --config ' +
-  path.join('.', 'node_modules', 'web-animations-set', 'src', '.eslintrc.json'),
-  { stdio: 'inherit' }
-).then(
+copyTestFiles()
+execa('node generate.js', { stdio: 'inherit' }).then(
   () => {
     execa(
-      path.join('.', 'node_modules', '.bin', 'prettier') +
-      ' --write "**/*.*" -u --print-width 280 --no-semi --single-quote --ignore-path ' +
-      path.join('.', 'node_modules', 'web-animations-set', 'src', '.prettierignore'),
+      path.join('.', 'node_modules', '.bin', 'eslint') +
+      ' --fix "**/*.*" --resolve-plugins-relative-to . --ignore-path ' +
+      path.join('.', 'node_modules', 'web-animations-set', 'src', '.eslintignore') +
+      ' --config ' +
+      path.join('.', 'node_modules', 'web-animations-set', 'src', '.eslintrc.json'),
       { stdio: 'inherit' }
     ).then(
       () => {
-        copyTestFiles()
-        execa(path.join('.', 'node_modules', '.bin', 'jest'), { stdio: 'inherit' }).then(
+        execa(
+          path.join('.', 'node_modules', '.bin', 'prettier') +
+          ' --write "**/*.*" -u --print-width 280 --no-semi --single-quote --ignore-path ' +
+          path.join('.', 'node_modules', 'web-animations-set', 'src', '.prettierignore'),
+          { stdio: 'inherit' }
+        ).then(
           () => {
-            execa(path.join('.', 'node_modules', '.bin', 'rollup -c'), { stdio: 'inherit' }).then(
+            execa(path.join('.', 'node_modules', '.bin', 'jest'), { stdio: 'inherit' }).then(
               () => {
-                execa(path.join('.', 'node_modules', '.bin', 'rollup -c rollup.config.split.js'), { stdio: 'inherit' }).then(
+                execa(path.join('.', 'node_modules', '.bin', 'rollup -c'), { stdio: 'inherit' }).then(
                   () => {
-                    removeTestFiles()
+                    execa(path.join('.', 'node_modules', '.bin', 'rollup -c rollup.config.split.js'), { stdio: 'inherit' }).then(
+                      () => {
+                        onError()
+                      },
+                      () => {
+                        onError()
+                      }
+                    )
                   },
                   () => {
-                    removeTestFiles()
+                    onError()
                   }
                 )
               },
               () => {
-                removeTestFiles()
+                onError()
               }
             )
           },
-          () => {
-            removeTestFiles()
+          (error) => {
+            onError(error)
           }
         )
       },
@@ -64,7 +75,7 @@ execa(
       }
     )
   },
-  (error) => {
-    onError(error)
+  () => {
+    onError()
   }
 )
